push to master

+++ lms_frontend/src/views/admin/course_admin/CourseDetail.vue
@@ -152,16 +152,20 @@
 
             <div v-for="(item, i) in items" :key="i">
               <v-hover v-slot:default="{ hover }" open-delay="80">
-                <v-card
-                  outlined
-                  max-width="800"
-                  class="mx-auto mt-6 rounded-lg border"
-                  style="border: 2px solid pink !important"
-                  :elevation="hover ? 8 : 0"
-                  min-height="150"
+                <router-link
+                  class="text-decoration-none"
+                  :to="'detail/' + item.uuid + '&' + title"
                 >
-                  <!-- Delete -->
-                  <v-dialog v-model="dialog3" max-width="290">
+                  <v-card
+                    outlined
+                    max-width="800"
+                    class="mx-auto mt-6 rounded-lg border"
+                    style="border: 2px solid pink !important"
+                    :elevation="hover ? 8 : 0"
+                    min-height="150"
+                  >
+                    <!-- Delete -->
+                    <!-- <v-dialog v-model="dialog3" max-width="290">
                     <v-card>
                       <v-card-title class="headline"
                         >Are you sure?</v-card-title
@@ -188,12 +192,9 @@
                         </v-btn>
                       </v-card-actions>
                     </v-card>
-                  </v-dialog>
-                  <!-- End Delete Dialog -->
-                  <router-link
-                    class="text-decoration-none"
-                    :to="'detail/' + item.uuid + '&' + title"
-                  >
+                  </v-dialog> -->
+                    <!-- End Delete Dialog -->
+
                     <v-list-item>
                       <v-avatar color="pink" size="40" class="mr-3"
                         ><i
@@ -217,32 +218,32 @@
                         style="color: red"
                         @click="dialog3 = true"
                       ></i> -->
-                      <v-btn color="black" style="float: right" icon>
+                      <v-btn
+                        color="black"
+                        style="float: right"
+                        icon
+                        @click="deleteexam(item)"
+                      >
                         <i
                           class="fa fa-trash float-right mx-1 align-top"
                           style="color: red"
-                          @click="dialog3 = true"
                         ></i>
                       </v-btn>
-
-                      <!-- <span class="float-right">{{
-                        convertDate(item.created_at)
-                      }}</span> -->
                     </v-list-item>
                     <v-divider class="my-2"></v-divider>
 
                     <v-card-text class="black--text">
                       {{ item.description }}
                     </v-card-text>
-                  </router-link>
 
-                  <!-- <i
+                    <!-- <i
                     class="far fa-edit float-right"
                     style="color: green"
                     color="primary"
                     @click="onEditItem(item)"
                   ></i> -->
-                </v-card>
+                  </v-card>
+                </router-link>
               </v-hover>
             </div>
           </v-tab-item>
@@ -360,8 +361,11 @@ export default {
     async deleteexam(item) {
       const index = this.items.indexOf(item);
       this.deletItems = item;
-      this.items.splice(index, 1);
-      await this.deleteExam();
+      const btnCancel = confirm("Are you sure you want to delete this item?");
+      if (btnCancel === true) {
+        this.items.splice(index, 1);
+        await this.deleteExam();
+      }
     },
     async deleteExam() {
       new Promise((resolve) => {
